backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
python-dotenv==1.0.1
groq==0.11.0
deepgram-sdk==3.7.0
cartesia==1.3.0
httpx>=0.27.2
